# Manually unroll x^7 for poseidon for a ~50% speedup [master]