Cherry-picked https://github.com/o1-labs/proof-systems/pull/2394.

Output on my machine:

```
poseidon_absorb_permutation kimchi                                                                             
                        time:   [30.086 µs 30.131 µs 30.221 µs]
                        change: [-40.050% -39.774% -39.371%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  10 (10.00%) high severe

poseidon_absorb_permutation legacy                                                                             
                        time:   [49.133 µs 49.157 µs 49.180 µs]
                        change: [-7.0440% -6.9919% -6.9411%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low severe
  5 (5.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe
```

Therefore getting ~40% improvement compared to current version.